# mall_dash_mobile/openapi_latest.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MalDash.API | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://maldash-development-api.runasp.net/"
    }
  ],
  "paths": {
...
    "/MalDashApi/Audit": {
      "get": {
        "tags": [
          "Audit"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          },
          {
            "name": "entityName",
            "in": "query",
            "schema": {
              "type": "string",
              "default": null
            }
          },
          {
            "name": "entityId",
            "in": "query",
            "schema": {
              "type": "string",
              "default": null
            }
          },
          {
            "name": "userId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid",
              "default": null
            }
          },
          {
            "name": "action",
            "in": "query",
            "schema": {
              "type": "string",
              "default": null
            }
          },
          {
            "name": "fromDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time",
              "default": null
            }
          },
          {
            "name": "toDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time",
              "default": null
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/MalDashApi/Audit/{id}": {
      "get": {
        "tags": [
          "Audit"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/MalDashApi/Audit/history/{entityName}/{entityId}": {
      "get": {
        "tags": [
          "Audit"
        ],
        "parameters": [
          {
            "name": "entityName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entityId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/MalDashApi/Building/{id}": {
      "get": {
        "tags": [
          "Building"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Building"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/MalDashApi/Building": {
      "get": {
        "tags": [
          "Building"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "searchName",
            "in": "query",
            "schema": {
              "type": "string",
              "default": null
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Building"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBuildingRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBuildingRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBuildingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/MalDashApi/Building/BuildingName/{id}": {
      "put": {
        "tags": [
          "Building"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBuildingRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBuildingRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBuildingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/MalDashApi/Building/{id}/floor": {
      "post": {
        "tags": [
          "Building"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/MalDashApi/Building/floor/{floorId}": {
      "delete": {
        "tags": [
          "Building"
        ],
        "parameters": [
          {
            "name": "floorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/MalDashApi/Building/floor/{floorId}/apartment": {
      "post": {
        "tags": [
          "Building"
        ],
        "parameters": [
          {
            "name": "floorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddApartmentRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddApartmentRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddApartmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/MalDashApi/Building/apartment/{apartmentId}": {
      "put": {
        "tags": [
          "Building"
        ],
        "parameters": [
          {
            "name": "apartmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateApartmentRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateApartmentRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateApartmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Building"
        ],
        "parameters": [
          {
            "name": "apartmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddApartmentRequest": {
        "required": [
          "apartmentName"
        ],
        "type": "object",
        "properties": {
          "apartmentName": {
            "type": "string"
          }
        }
      },
      "ApartmentLayoutRequest": {
        "type": "object",
        "properties": {
          "rooms": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoomLayoutRequest"
            }
          },
          "doors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DoorRequest"
            }
          },
          "gridSize": {
            "type": "integer",
            "format": "int32"
          }
        },
        "nullable": true
      },
      "CreateBuildingRequest": {
        "required": [
          "name",
          "floors"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "floors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FloorRequest"
            }
          }
        }
      },
...
      "DoorRequest": {
        "required": [
          "id",
          "roomId",
          "edge",
          "position",
          "width"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "roomId": {
            "type": "string"
          },
          "connectedRoomId": {
            "type": "string",
            "nullable": true
          },
          "edge": {
            "type": "string"
          },
          "position": {
            "type": "number",
            "format": "double"
          },
          "width": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "FloorRequest": {
        "required": [
          "floorNumber",
          "numberOfApartments"
        ],
        "type": "object",
        "properties": {
          "floorNumber": {
            "type": "integer",
            "format": "int32"
          },
          "numberOfApartments": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "RoomLayoutRequest": {
        "required": [
          "id",
          "type",
          "name",
          "x",
          "y",
          "width",
          "height"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "x": {
            "type": "number",
            "format": "double"
          },
          "y": {
            "type": "number",
            "format": "double"
          },
          "width": {
            "type": "number",
            "format": "double"
          },
          "height": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "UpdateApartmentRequest": {
        "required": [
          "apartmentName"
        ],
        "type": "object",
        "properties": {
          "apartmentName": {
            "type": "string"
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "layout": {
            "$ref": "#/components/schemas/ApartmentLayoutRequest"
          }
        }
      },
      "UpdateBuildingRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}